refactor: AuthDIContainerとHealthDIContainerのLogger実装を共通化 HOXBL-94
- 両DIコンテナに重複していたinline Logger実装をshared/logging/ConsoleLogger.tsに集約
- getLogger()内の到達不能な例外分岐（デッドコード）を削除
